+++ cassandra/src/main/scala/geotrellis/store/cassandra/package.scala
@@ -24,7 +24,7 @@ import com.datastax.oss.driver.api.core.cql.{AsyncResultSet, Statement}
 
 import java.math.BigInteger
 
-package object cassandra extends java.io.Serializable {
+package object cassandra extends Serializable {
   implicit class BigIntOps(val self: BigInt) extends AnyVal {
     def asJava: BigInteger = new BigInteger(self.toByteArray)
   }